Adjust: Organizing compose files
