+++ app/src/main/java/dev/lhalegria/dogga/view/composable/BreedListItem.kt
@@ -1,0 +1,36 @@
+package dev.lhalegria.dogga.view.composable
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import dev.lhalegria.dogga.model.BreedModel
+
+@Composable
+fun BreedListItem(dog: BreedModel, clickAction: () -> Unit) {
+    Card(
+        modifier = Modifier
+            .padding(top = 8.dp, bottom = 4.dp, start = 16.dp, end = 16.dp)
+            .fillMaxWidth()
+            .wrapContentHeight(align = Alignment.Top)
+            .clickable { clickAction.invoke() },
+        elevation = CardDefaults.cardElevation(defaultElevation = 5.dp)
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.Start
+        ) {
+            BreedListItemPicture(dog.photo, 72.dp)
+            BreedListItemText(dog.name, Alignment.Start)
+        }
+    }
+}
