+++ app/src/main/java/dev/lhalegria/dogga/view/composable/Loading.kt
@@ -1,0 +1,24 @@
+package dev.lhalegria.dogga.view.composable
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import dev.lhalegria.dogga.view.ui.theme.caramel
+
+@Composable
+fun LoadingBox() {
+    Box(Modifier.fillMaxSize()) {
+        CircularProgressIndicator(
+            modifier = Modifier
+                .align(Alignment.Center)
+                .padding(50.dp),
+            progress = 1f,
+            color = caramel
+        )
+    }
+}
