+++ app/src/main/java/dev/lhalegria/dogga/view/ui/theme/Color.kt
@@ -11,3 +11,4 @@ val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
 
 val caramel = Color(0xFFA58D11)
+val caramelStrong = Color(0xFF756304)
