+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">Dogga</string>
+    <string name="empty_data_message">There are no data to show</string>
+    <string name="error_message">There was an error while trying to get the data</string>
+    <string name="try_again">Try again</string>
 </resources>
